+++ Source/OzzContextGen.WPF/MainWindow.xaml.cs
@@ -1,0 +1,17 @@
+﻿using OzzContextGen.WPF.ViewModels;
+using System.Windows;
+
+namespace OzzContextGen.WPF
+{
+    /// <summary>
+    /// Interaction logic for MainWindow.xaml
+    /// </summary>
+    public partial class MainWindow : Window
+    {
+        public MainWindow()
+        {
+            InitializeComponent();
+            this.DataContext = new MainViewModel();
+        }
+    }
+}